Extract operator to separate directory
