+++ .gitignore
@@ -14,4 +14,6 @@
 .glide/
 
 bank-vaults
-tmp/_output/bin/*
+
+# Operator
+operator/tmp/_output/bin/*
