+++ README.md
@@ -190,13 +190,15 @@ We have a fully fledged, production ready [Helm chart](https://github.com/banzai
 
 We have a Vault operator built on bank-vaults features as:
 
-- TLS support
 - external, API based configuration (secret engines, auth methods, policies) to automatically re/configure a Vault cluster
 - automatic unsealing (AWS, GCE, Azure, Kubernetes Secrets (for dev purposes), Oracle)
+- TLS support
 
 The operator flow is the following:
 
 ![operator](docs/images/vaultoperator.png)
+
+The source code can be found inside the [operator](operator/) directory.
 
 ### Deploying the operator
