+++ operator/README.md
@@ -1,0 +1,10 @@
+# Vault Operator
+
+This directory holds the code of the Banzai Cloud Vault Operator.
+
+## Build
+
+```bash
+cd operator
+operator-sdk build banzaicloud/vault-operator
+```
