+++ operator/cmd/vault-operator/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"runtime"
 
-	stub "github.com/banzaicloud/bank-vaults/pkg/stub"
+	stub "github.com/banzaicloud/bank-vaults/operator/pkg/stub"
 	sdk "github.com/operator-framework/operator-sdk/pkg/sdk"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
