+++ operator/deploy/cr-gcs-ha.yaml
