+++ operator/pkg/apis/vault/v1alpha1/register.go
