+++ operator/pkg/apis/vault/v1alpha1/types.go
